apachebench: create MAINTAINERS.html
Run the apachebench install script from the bench script. The install
script for apachebench does not run and the MAINTAINERS.html file is not
created in the htdocs directory, which leads to apachebench not being able
to detect a regression that is reproducible with a document of roughly 400
KiB in size. Additional clean up of the install script.

Signed-off-by: Jiri Wiesner <jwiesner@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>