netperf: Record interim results over time
netperf can report high variance and it can be difficult to debug as
only the final result based on a confidence interval is reported.
As the initial step in dealing with this, record interim results as
the test runs by recording /proc/net/dev over time and using the
netperf demo mode.

Signed-off-by: Mel Gorman <mgorman@suse.com>