netperf: Report filtered "interim" results
When compiled in demo mode, netperf can optionally display "interim" results
approximately every 1 second. Generate a report that filters the interim
results and display a new estimated report including how many results
were filtered. The intent is to help detect when variability is a major
problem. The filter is not statistically relevant or a better estimate
of the "true mean" and is only intended as a guide. The objective would
still be to identify what the source of the variability is and improve
the estimate of the true mean.

Signed-off-by: Mel Gorman <mgorman@suse.com>