run-container/S-startup: Add option for privileged run and use it for S-startup
S-startup requires several privileges to run properly in container
environment.

Performance numbers even in privileged mode are already much worse in
comparison to when it is run natively.

At this moment it does not make sense to add special handling for each
required privileged access. Instead lift all restrictions for this
benchmark. Cherry-picking privileges to try to run this benchmark in a
confined container with reduced privileges can be done if really
required at a later point.

Signed-off-by: Andreas Herrmann <aherrmann@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>