Summarise: Add support for specifying preferred value
Currently the preferred value is infered from data type specified in the
package. This is a bit cumbersome as e.g. measuring in kilobytes does
not quite say which direction is "better" so we have to create multiple
datatypes for "kilobytes" to express which is better. Provide a method
for a module to define preferred value if it does not define a data
type. The method is a bit weird to support both old and new scheme but
it gets trivial once we get rid of the old scheme.

Signed-off-by: Jan Kara <jack@suse.cz>