SummariseMonitor: Make monitors default to lower-is-better
For monitors it is difficult to define whether lower or higher values
are actually better because it depends on the workload the monitor is
watching. So just make monitors default to lower-is-better and see
whether something breaks or people get confused. We can always revisit
this or override the default for a particular monitor.

Signed-off-by: Jan Kara <jack@suse.cz>