Stat: Use proper mean for stddev calculation
Currently we always used arithmetic mean for calculation of standard
deviation. This is not matching the mean we use to compute for reporting
and also the ratio comparisons are mismatching (as the standard
deviation does not easily characterize the distribution of the harmonic
mean). So make standard deviation computation use harmonic mean if
appropriate and compute "harmonic standard deviation" (it corresponds to
the fact that we expect 1/X to be close to normally distributed random
variables). Also update computation of coeffvar accordingly and merge
code for computing parametrized functions in runStatFunc.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>