Summarise: Make submean calculations more standard
Currently submean calculations were rather special because they also
computed confidence interval length to save CPU time and returned tuple
instead of a single value. This needed some specialcasing in the code.
Now when we cache results of statistical functions in $stats hash, we
can just save all computed statistics there and avoid most of the
specialcasing. We also rename the submean functions to submean-?mean and
submeanci-?mean to make it consistent with other parametrized
statistical functions. Finally, we also provide low and high boundary of
the mean confidence interval as eventually we will use that for
comparisons.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>