Compare: Use CI interval for significance testing
Instead of doing Student's t-test for evaluating statistical
significance of result difference, use computed confidence intervals.
They are somewhat statistically weaker but apply for all the cases where
we can compute the confidence interval of a given statistical property
(such as mean, percentile, or so).

We deliberately use RatioSummaryStats for determining the entity of
interest for comparison and its confidence interval, so that the
significance results in ratio & normal comparisons are the same.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>