configs/*kernbench*: Use stable kernel source tree for compilation
The default v5.14 does not compile (neither defconfig nor
allmodconfig) with GCC14 which is used on SLE16.

Attaching patches to the shellpack to fix corresponding issues is not
an option because it would have to be kernel version specific.

Instead selecting a long term stable kernel version wich already
contains fixes for compiler issues is the optimal solution.
I've decided to use v5.15.178.

I've double checked that this kernel successfully compiles under
SLE15SP6 and SLE16 and thus can be used for this compile benchmark.

Signed-off-by: Andreas Herrmann <aherrmann@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>