netperf: Update to 2.6.0
Our QE team runs this version. It is also getting increasingly dificult
to build 2.4.5 with newer gcc.  So bite the bullet and update.

This release requires an out-of-tree patch to build. The build issue existed
upstream for two releases and is now fixed in master but unreleased.  We
cannot backport the upstream patch because it is a gigantic mess fixing many
things and wouldn't apply. Instead, this adds a trivial two line fix
only for this build issue.

QE doesn't hit the build issue because they build with --enable-demo=no.

Signed-off-by: Gabriel Krisman Bertazi <krisman@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>